test delete user

+++ src/main/java/com/example/controller/UserController.java
@@ -38,6 +38,10 @@ public class UserController {
 
     @DeleteMapping("/delete/{userId}")
     public String deleteUserById(@PathVariable UUID userId) {
+        User user = userService.getUserById(userId);
+        if(user == null){
+            return "User not found";
+        }
         userService.deleteUserById(userId);
         return "User deleted successfully";
     }
@@ -65,7 +69,7 @@ public class UserController {
     @PutMapping("/addProductToCart")
     public String addProductToCart(@RequestParam UUID userId, @RequestParam UUID productId) {
         userService.addProductToCart(userId, productId);
-        return "Product with ID " + productId + " added to user " + userId + "'s cart.";
+        return "Product added to cart";
     }
 
     @PutMapping("/deleteProductFromCart")
